requirements.txt
flask==2.3.2
pyyaml==6.0
gunicorn==20.1.0